requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-appconf==1.0.4
django-choices==1.7.1
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-font-icons==1.0.1
django-geoip==0.5.2.1
django-js-asset==1.2.2
googlemaps==4.4.2
idna==2.10
Pillow==8.1.2
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
twilio==6.53.0
urllib3==1.26.3
